Data changes

+++ app.js
@@ -1,6 +1,7 @@
-var ftp = require('ftp-get')
-var fs  = require('fs');
-var csv = require("csvtojson").Converter;
+var ftp     = require('ftp-get')
+var fs      = require('fs');
+var csv     = require("csvtojson").Converter;
+var request = require("request");
 
 var remoteFile = "ftp://ftp.nasdaqtrader.com/SymbolDirectory/nasdaqtraded.txt"
 var timestamp  = String((new Date).getTime());
@@ -8,12 +9,21 @@ var filePath   = "./data/nasdaq/"
 var symbolPath = "./data/symbols/"
 var localFile  = ""
 var extension  = ".psv"
+var keys       = [];
+var current    = 0;
 
-function makeFolder(dir, callback) {
+function makeFolder(dir) {
 	if (!fs.existsSync(dir)){
 	    fs.mkdirSync(dir);
 	}
-	callback();
+}
+
+function fileExists(filePath) {
+    try {
+        return fs.statSync(filePath).isFile();
+    } catch (err) {
+        return false;
+    }
 }
 
 function getLatestSymbols(callback) {
@@ -50,29 +60,63 @@ function pad(width, string, padding) {
 }
 
 function readLatestSymbols(latest) {
-	var converter = new csv({
-		delimiter:"|"
-	});
-	converter.on("end_parsed", function (jsonArray) {
-		var obj = {}
-		for (i in jsonArray) {
-			symbol = jsonArray[i]["Symbol"];
-			security = jsonArray[i]["Security Name"];
-			obj[symbol] = security
-			console.log(pad(6, symbol, " ") + " : " + security);
-		}
-		// console.log(obj);
-	    fs.writeFile(symbolPath + timestamp + extension, JSON.stringify(obj), function (err) {
-		  	if (err) return console.log(err);
+	if (!fileExists(symbolPath + latest + extension)) {
+		console.log("Generating New Symbol Data")
+		var converter = new csv({
+			delimiter:"|"
 		});
-	}); 
-	fs.createReadStream(filePath + String(latest) + extension).pipe(converter);
+		converter.on("end_parsed", function (jsonArray) {
+			var obj = {}
+			for (i in jsonArray) {
+				symbol = jsonArray[i]["Symbol"];
+				security = jsonArray[i]["Security Name"];
+				obj[symbol] = security
+				console.log(pad(6, symbol, " ") + " : " + security);
+			}
+			// console.log(obj);
+		    fs.writeFile(symbolPath + timestamp + extension, JSON.stringify(obj), function (err) {
+			  	if (err) return console.log(err);
+			});
+		}); 
+		fs.createReadStream(filePath + String(latest) + extension).pipe(converter);
+		readLatestSymbols(latest);
+	} else {
+		console.log("Using Cached Version of Symbol Data");
+		listSymbolData(latest);
+	}
 }
 
-makeFolder("./data", function(){
-	makeFolder("./data/nasdaq", function() {
-		makeFolder("./data/symbols", function() {
-			checkLatestSymbols();
-		});
+function listSymbolData(latest) {
+	fs.readFile(symbolPath + latest + extension, function(err, data) {
+		if (err) throw err;
+		var symbolList = JSON.parse(data);
+		keys = Object.keys(symbolList);
+		symbolCall();
 	});
-});
+}
+
+function symbolCall() {
+	getSymbolData(keys[current], function(symbol, body) {
+		current += 1;
+		sortData(symbol, body);
+		symbolCall();
+	});
+}
+
+function getSymbolData(symbol, callback) {
+	var url = 'http://dev.markitondemand.com/MODApis/Api/v2/InteractiveChart/json?parameters={"Normalized":false,"NumberOfDays":365,"DataPeriod":"Day","Elements":[{"Symbol":"' + symbol + '","Type":"price","Params":["c"]}]}';
+	request({
+	  	uri: url,
+	}, function(error, response, body) {
+			callback(symbol, body);
+	});
+}
+
+function sortData(symbol, body) {
+	console.log(body);
+}
+
+makeFolder("./data");
+makeFolder("./data/nasdaq");
+makeFolder("./data/symbols");
+checkLatestSymbols();
